Unified Service tests per action
