--- Example/Tests/KWSKWSRequestPermissionTests.m
@@ -1,27 +1,0 @@
-//
-//  KWSKWSRequestPermissionTests.m
-//  KWSiOSSDKObjC
-//
-//  Created by Gabriel Coman on 27/05/2016.
-//  Copyright © 2016 Gabriel Coman. All rights reserved.
-//
-
-#import <XCTest/XCTest.h>
-
-@interface KWSKWSRequestPermissionTests : XCTestCase
-
-@end
-
-@implementation KWSKWSRequestPermissionTests
-
-- (void)setUp {
-    [super setUp];
-    // Put setup code here. This method is called before the invocation of each test method in the class.
-}
-
-- (void)tearDown {
-    // Put teardown code here. This method is called after the invocation of each test method in the class.
-    [super tearDown];
-}
-
-@end
