+++ Example/Tests/Leaders+ModelTests.swift
@@ -1,5 +1,5 @@
 //
-//  LeadersTests.swift
+//  Leaders+ModelTests.swift
 //  KWSiOSSDKObjC_Tests
 //
 //  Created by Guilherme Mota on 10/04/2018.
@@ -10,7 +10,7 @@ import XCTest
 import Nimble
 import KWSiOSSDKObjC
 
-class LeadersTests: XCTestCase {
+class LeadersModelTests: XCTestCase {
     
     override func setUp() {
         super.setUp()
