+++ Example/Tests/Scoring+ServiceTests.swift
@@ -1,8 +1,8 @@
 //
-//  GetLeaders+ServiceTests.swift
+//  Scoring+ServiceTests.swift
 //  KWSiOSSDKObjC_Tests
 //
-//  Created by Guilherme Mota on 10/04/2018.
+//  Created by Guilherme Mota on 19/04/2018.
 //  Copyright © 2018 Gabriel Coman. All rights reserved.
 //
 
@@ -13,7 +13,7 @@ import KWSiOSSDKObjC
 import SAMobileBase
 import SAProtobufs
 
-class GetLeadersServiceTests: XCTestCase {
+class ScoringServiceTests: XCTestCase {
     
     //class or data to test
     private var scoringService: ScoringServiceProtocol!
@@ -40,6 +40,78 @@ class GetLeadersServiceTests: XCTestCase {
         scoringService = nil
         environment = nil
     }
+    
+    //MARK: GET SCORE
+    
+    func test_Scoring_GetScore_ValidRequestAndResponse(){
+        
+        let JSON: Any? = try? fixtureWithName(name:"get_user_score_success_response")
+        
+        //when
+        stub(everything , json(JSON!))
+        
+        waitUntil { done in
+            
+            self.scoringService.getScore(appId: self.goodAppId,
+                                         token: self.goodToken,
+                                         completionHandler:  { (score, error) in
+                                            
+                                            expect(score).toNot(beNil())
+                                            expect(score?.score).to(equal(600))
+                                            expect(score?.rank).to(equal(1))
+                                            expect(error).to(beNil())
+                                            
+                                            done()
+            })
+        }
+    }
+    
+    func test_Scoring_GetScore_BadToken_Response(){
+        let JSON: Any? = try? fixtureWithName(name:"generic_invalid_token_response")
+        
+        //when
+        stub(everything , json(JSON!, status: 401))
+        
+        waitUntil { done in
+            
+            self.scoringService.getScore(appId: self.goodAppId,
+                                         token: self.badToken,
+                                         completionHandler:  { (score, error) in
+                                            
+                                            expect(score).to(beNil())
+                                            expect(error).toNot(beNil())
+                                            expect((error as! ErrorWrapper).errorCode).to(equal("invalid_token"))
+                                            expect((error as! ErrorWrapper).error).to(equal("The access token provided is invalid."))
+                                            
+                                            done()
+            })
+        }
+    }
+    
+    func test_Scoring_GetScore_BadClientId_Response(){
+        let JSON: Any? = try? fixtureWithName(name:"generic_operation_not_supported_for_client_response")
+        
+        //when
+        stub(everything , json(JSON!, status: 403))
+        
+        waitUntil { done in
+            
+            self.scoringService.getScore(appId: self.badAppId,
+                                         token: self.goodToken,
+                                         completionHandler:  { (score, error) in
+                                            
+                                            expect(score).to(beNil())
+                                            expect(error).toNot(beNil())
+                                            expect((error as! ErrorWrapper).code).to(equal(1))
+                                            expect((error as! ErrorWrapper).codeMeaning).to(equal("forbidden"))
+                                            expect((error as! ErrorWrapper).message).to(equal("operation not supported for this client"))
+                                            
+                                            done()
+            })
+        }
+    }
+    
+    //MARK: GET LEADERS
     
     func test_Scoring_GetLeaders_ValidRequestAndResponse(){
         
@@ -75,7 +147,7 @@ class GetLeadersServiceTests: XCTestCase {
         }
     }
     
-    func test_Scoring_GetLeaders_BadTokenResponse(){
+    func test_Scoring_GetLeaders_BadToken_Response(){
         let JSON: Any? = try? fixtureWithName(name:"generic_invalid_token_response")
         
         //when
@@ -85,7 +157,7 @@ class GetLeadersServiceTests: XCTestCase {
             
             self.scoringService.getLeaderboard(appId: self.goodAppId,
                                                token: self.badToken,
-                completionHandler:  { (leader, error) in
+                                               completionHandler:  { (leader, error) in
                                                 
                                                 expect(error).toNot(beNil())
                                                 expect((error as! ErrorWrapper).errorCode).to(equal("invalid_token"))
