--- Example/Tests/Tests.m
@@ -1,34 +1,0 @@
-//
-//  KWSiOSSDKObjCTests.m
-//  KWSiOSSDKObjCTests
-//
-//  Created by Gabriel Coman on 05/24/2016.
-//  Copyright (c) 2016 Gabriel Coman. All rights reserved.
-//
-
-@import XCTest;
-
-@interface Tests : XCTestCase
-
-@end
-
-@implementation Tests
-
-- (void)setUp
-{
-    [super setUp];
-    // Put setup code here. This method is called before the invocation of each test method in the class.
-}
-
-- (void)tearDown
-{
-    // Put teardown code here. This method is called after the invocation of each test method in the class.
-    [super tearDown];
-}
-
-- (void)testExample {
-    
-}
-
-@end
-
